+++ bajiecloud-product-server/src/main/java/com/bajiezu/cloud/product/controller/vo/response/AssetResidualYearConfigVO.java
@@ -10,9 +10,9 @@ public class AssetResidualYearConfigVO {
     private Integer useYear;
     private BigDecimal totalPriceUpperCoefficient;
     private BigDecimal totalPriceLowerCoefficient;
-    private Long yearBeginValue;
-    private Long yearDepreciationAmount;
-    private Long yearEndResidualValue;
-    private Long totalPriceUpperLimit;
-    private Long totalPriceLowerLimit;
+    private BigDecimal yearBeginValue;
+    private BigDecimal yearDepreciationAmount;
+    private BigDecimal yearEndResidualValue;
+    private BigDecimal totalPriceUpperLimit;
+    private BigDecimal totalPriceLowerLimit;
 }
